requirements.txt
aiofile
asyncpg
fastapi
httpx
python-jose[cryptography]
pyyaml
SQLAlchemy
uvicorn